removed idaes.util.system and idaes.util, and refactored accordingly
