+++ idaes/core/dmf/tests/test_commands.py
@@ -23,7 +23,7 @@ import pytest
 
 #
 from idaes.core.dmf import dmfbase, commands, errors, workspace, util
-from idaes.util.system import mkdtemp
+from idaes.core.dmf.util import mkdtemp
 from .util import init_logging
 
 __author__ = "Dan Gunter"
